greatly improve eraser performance
Load uploaded image to a background canvas, then draw on top of it.
That way, the erase option is just a clearRect on the background
canvas.

At save time, draw the background canvas and foreground (drawing)
canvas to an offscreen canvas and open the resulting image in a new
window.